# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/workdays/{workdayId}/corrections:
    post:
      tags:
      - Corrections
      operationId: request
      parameters:
      - name: workdayId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionRequestDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorrectionResponse'
...
  /api/v1/public/tenant-registrations:
    post:
      tags:
      - Public - Tenant registrations
      summary: Solicita el alta de una organización (RF-REG-001)
      operationId: request_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantRegistrationRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TenantRegistrationAcceptedResponse'
...
  /api/v1/platform/registrations/{registrationId}/approve:
    post:
      tags:
      - Platform - Tenant registrations
      summary: Aprueba una solicitud y crea el tenant en estado PENDING
      operationId: approve
      parameters:
      - name: registrationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TenantRegistrationResponse'
  /api/v1/notifications/{notificationId}/read:
    post:
      tags:
      - Notifications
      summary: Marca como leida una notificacion propia
      operationId: markRead
      parameters:
      - name: notificationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/corrections/{correctionId}/approve:
    post:
      tags:
      - Corrections
      operationId: approve_1
      parameters:
      - name: correctionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionResolutionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorrectionResponse'
...
  /api/v1/app/absences:
    get:
      tags:
      - Absences
      operationId: list_2
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbsenceResponse'
    post:
      tags:
      - Absences
      operationId: request_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbsenceRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AbsenceResponse'
...
  /api/v1/admin/absences/{absenceId}/reject:
    post:
      tags:
      - Admin - Absences
      operationId: reject_2
      parameters:
      - name: absenceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbsenceResolutionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AbsenceResponse'
  /api/v1/admin/absences/{absenceId}/approve:
    post:
      tags:
      - Admin - Absences
      operationId: approve_2
      parameters:
      - name: absenceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbsenceResolutionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AbsenceResponse'
...
  /api/v1/reports/tenant/summary:
    get:
      tags:
      - Reports
      operationId: tenantSummary
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantEmployeeSummaryResponse'
  /api/v1/reports/tenant/export.pdf:
    get:
      tags:
      - Reports
      summary: Exporta el informe agregado del tenant en PDF
      operationId: tenantSummaryPdf
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
  /api/v1/reports/tenant/export.csv:
    get:
      tags:
      - Reports
      operationId: tenantSummaryCsv
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/reports/employees/{employeeId}/summary:
    get:
      tags:
      - Reports
      operationId: employeeSummary
      parameters:
      - name: employeeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeDaySummaryResponse'
...
  /api/v1/notifications:
    get:
      tags:
      - Notifications
      summary: Lista las notificaciones del usuario autenticado
      operationId: list_7
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedNotificationsResponse'
...
  /api/v1/admin/absences:
    get:
      tags:
      - Admin - Absences
      operationId: list_13
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AbsenceResponse'
...
components:
  schemas:
...
    CorrectionRequestDto:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
          minLength: 0
        proposedChanges:
          $ref: '#/components/schemas/ProposedChangesDto'
      required:
      - proposedChanges
    ProposedBreakDto:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
      required:
      - endedAt
      - startedAt
    ProposedChangesDto:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        breaks:
          type: array
          items:
            $ref: '#/components/schemas/ProposedBreakDto'
          maxItems: 24
          minItems: 0
      required:
      - breaks
      - endedAt
      - startedAt
    CorrectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workdayId:
          type: string
          format: uuid
        requestedBy:
          type: string
          format: uuid
        reason:
          type: string
        proposedChanges:
          $ref: '#/components/schemas/ProposedChangesResponse'
        status:
          type: string
        resolvedBy:
          type: string
          format: uuid
        resolvedAt:
          type: string
          format: date-time
        resolutionComment:
          type: string
        createdAt:
          type: string
          format: date-time
    ProposedBreakResponse:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
    ProposedChangesResponse:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        breaks:
          type: array
          items:
            $ref: '#/components/schemas/ProposedBreakResponse'
...
    TenantRegistrationRequestBody:
      type: object
      properties:
        companyName:
          type: string
          maxLength: 200
          minLength: 0
        timezone:
          type: string
          minLength: 1
        firstName:
          type: string
          maxLength: 200
          minLength: 0
        lastName:
          type: string
          maxLength: 200
          minLength: 0
        email:
          type: string
          maxLength: 255
          minLength: 0
        password:
          type: string
          maxLength: 200
          minLength: 10
        acceptTerms:
          type: boolean
    TenantRegistrationAcceptedResponse:
      type: object
      properties:
        message:
          type: string
...
    TenantRegistrationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
        ownerFirstName:
          type: string
        ownerLastName:
          type: string
        email:
          type: string
        timezone:
          type: string
        status:
          type: string
        source:
          type: string
        decisionReason:
          type: string
        createdTenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
        decidedAt:
          type: string
          format: date-time
...
    CorrectionResolutionRequest:
      type: object
      properties:
        resolutionComment:
          type: string
          maxLength: 500
          minLength: 0
...
    AbsenceRequestBody:
      type: object
      properties:
        absenceTypeId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
          maxLength: 500
          minLength: 0
      required:
      - absenceTypeId
      - endDate
      - startDate
    AbsenceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        absenceTypeId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        reason:
          type: string
        status:
          type: string
        resolvedBy:
          type: string
          format: uuid
        resolvedAt:
          type: string
          format: date-time
        resolutionComment:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AbsenceResolutionRequest:
      type: object
      properties:
        resolutionComment:
          type: string
          maxLength: 500
          minLength: 0
...
    TenantEmployeeSummaryResponse:
      type: object
      properties:
        employeeId:
          type: string
          format: uuid
        worked:
          type: string
          example: PT2H30M
        paused:
          type: string
          example: PT2H30M
        expected:
          type: string
          example: PT2H30M
        effectiveWorked:
          type: string
          example: PT2H30M
        overtime:
          type: string
          example: PT2H30M
        deviation:
          type: string
          example: PT2H30M
        workdayCount:
          type: integer
          format: int32
        adjustedWorkdayCount:
          type: integer
          format: int32
        openWorkdays:
          type: integer
          format: int32
        evaluatedWorkdayCount:
          type: integer
          format: int32
    EmployeeDaySummaryResponse:
      type: object
      properties:
        day:
          type: string
          format: date
        worked:
          type: string
          example: PT2H30M
        paused:
          type: string
          example: PT2H30M
        expected:
          type: string
          example: PT2H30M
        effectiveWorked:
          type: string
          example: PT2H30M
        overtime:
          type: string
          example: PT2H30M
        deviation:
          type: string
          example: PT2H30M
        workdayCount:
          type: integer
          format: int32
        adjustedWorkdayCount:
          type: integer
          format: int32
        openWorkdays:
          type: integer
          format: int32
        evaluatedWorkdayCount:
          type: integer
          format: int32
...
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        read:
          type: boolean
    PagedNotificationsResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...